final touchups

+++ client/src/components/WorkoutDeatils.jsx
@@ -1,13 +1,20 @@
-/* eslint-disable react/prop-types */
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import { formatDistance } from "date-fns";
 import { MdAutoDelete } from "react-icons/md";
 
 const WorkoutDeatils = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
     if (response.ok) {
